Added ifdef for arduMSP or fitstatUSB for device selection for debuging, added…

See merge request Andrew/fit-statUSB!14

+++ myTimers.h
@@ -9,7 +9,7 @@
 #define MYTIMERS_H_
 
 //float fadeTimer;
-int fadeTimer;
+long fadeTimer;
 uint32_t colortick;
 int timeCounter;
